Lift utility class for sharing
